Adding RewriteDependencyCheckPlugin to the base plugin
Some repositories are failing to complete the vulnerability check because they are missing the "DependencyCheck" which is provided by the DependencyCheckPlugin. We want to validate vulnerabilities on all repositiories but not all repositories have the "RewriteDependencyCheckPlugin" which applies the "DependencyCheckPlugin". Instead of adding it to all repositories, I'm adding this to the base plugin

Note, we are applying the RewriteDependencyCheckPlugin instead of the DependencyCheckPlugin to the base plugin. This is done because the RewriteDependencyCheckPlugin will apply the shared suppressions to to all repositories when we run the validation